Add metrics example
We often get questions from people asking about how to track metrics but
I've been reluctant to add an example for it since it required writing a
`tower::Service` middleware. However now that we have
`middleware::from_fn` I think it makes sense to have!

The longterm goal is for tower-http to provide something but I haven't
found the right design for it yet.