fix: warnings for cloning references
In #1674, it was pointed out that `axum-macros` causes some `clippy`
warnings (and some `rustc` lints to fail) as the `FromRef` macro
generates code which clones reference types unnecessarily.

This seems to be when generating the body of the trait implementation,
but since we have the type information from `syn` at that point it's
relatively simple to resolve these.

This change:
* Avoids using `.clone` if we are generating the implementation for a
  reference type field
* Adds a simple UI test to ensure this compiles with the `rustc` lint
  set to deny

Fixes #1674.