Require `StreamBody` to be infallible, add `TryStreamBody`
I've been thinking some more about #1699 and I think this approach makes
sense.

This changes `StreamBody` to always require the underlying stream to be
infallible, i.e. it must yield `impl Into<Bytes>`. `Result`s are not
allowed. That is because hyper will silently ignore body errors, so its
kind of a footgun.

Additionally I think we can provide a `TryStreamBody` that allows
`Result`s but also a callback to call on errors. Then we can use
`TryStreamBody` in `JsonLines` and `AsyncReadBody` which don't have a
good way to surface errors otherwise.