add `has_routes` function and docu to `Router`
The motivation is to be able to avoid panics in generic code that
wants to call `route_layer`.

For example, you might have a generic function that takes a `Router`
configured by an application (`/api/v1/my_route`, `/api/v1/my_route2`,...)
and adds a bunch of additional configuration that is common to your
applications, like an http metrics layer using `route_layer`, followed
by infrastructure or diagnostic routes (`/health`, `/ready`, ...)
that you do not need metrics for.

`route_layer` currently panics if you call it on `Router::new()`.
But in some cases that is legitimate, for example, you may have an
application that performs important periodic tasks, and needs to have
`/health` and `/ready` to exist in your k8s infrastructure,
but doesn't actualy have any HTTP API to expose. So then it would be
very logical to pass `Router::new()` to your generic function.

With this non-breaking change, the helper function would be able to
call `has_routes()` first on the argument, and skip the (panicking)
call to `route_layer` if the router has no routes. This is a bit nicer
than the workaround of adding dummy routes to the `Router::new()`
just to avoid a `Router` panic.