Clearer graceful shutdown doc example.
The current doc for with_graceful_shutdown() implies a select! block
somewhere else that races the serve and the cancellation futures. I
think using a `tokio::sync::Notify` makes for a more generally useful
pattern since it can be passed around and triggered as needed. For
example, the select! strategy is difficult to use in integration tests.