Remove broken `rest-grpc-multiplex` example
This example has not been updated to hyper v1 for a year and requires every developer to have `protoc` installed on their machine to be able to run `cargo build --workspace`. Both of these points combined IMHO shift the tradeoff enough to it not being valuable enough to keep the broken example around.