Generalize `Connected` implementation to all `Listener`s
The previous implementation implemented the trait `Connected` for `TcpListener`
and `TapIo`. This was both needlessly specific *and* blocked other implementers
of the `Listener` trait from themselves implementing `Connected`.

For example when using axum with
`into_make_service_with_connect_info::<SocketAddr>()` a `Listener`
implementation that isn't `TcpListener` will yield a compiler error that
`Connected` is not implemented for this other `Listener`. Implementing it for
`SocketAddr` directly isn't possible because of the orphan rule. Even with a
wrapper type, a library abstraction can't implement it generically for all
`Listener`s because it would run into conflicting trait implementations.