fix: explicitly match with `()` instead of `_`
Using `()` makes it clearer to both the reader and the compiler that the
pattern contains no data. It will also trigger a compilation error if
the type changes, which makes the pattern-matching more robust, safe,
and less prone to uncaught logic errors. This fixes the
`ignored_unit_patterns` clippy lint.