chore: centralize `LICENSE` to the repo root
Currently, the project `LICENSE` file exists in each crate's directory
only. This causes problems since GitHub is unable to figure out which
license the project uses [1].

This repo also includes multiple different variations of the same file
(different years, different wrapping, different symbols, etc). Simply
fixing each file to include the same content is not enough, as one
`LICENSE` file may be forgotten about in the future.

The solution? Centralizing the `LICENSE` file in the repo root and then
using symlinks in each crate to point to the `LICENSE` file allows us to
only have to change one file when updating the `LICENSE`. This helps
follow the principle of DRY.

I used `2019–2025` as the copyright year for this centralized `LICENSE`
file, since the year needed updating as well. Note that the dash is an
en dash (`U+2013`), which requires working Unicode support to display
correctly.

For the word wrapping, I used the original from when the repo was
originally set up [2].

[1]: https://github.com/tokio-rs/axum/community
[2]: 002e3f92b322fd6a9241d1c6fed36fc15604ef28