chore: remove useless `mut` in function parameters
The `oneshot_inner` method takes `self` as a mutable reference, when it
could just instead take `self` as a plain reference. Using `mut` is
confusing and unnecessary, and also means less fights with the borrow
checker if `oneshot_inner` is used in more places in the future. This
fixes the `needless_pass_by_ref_mut` clippy lint.