fix(axum-extra): don't require `S` generic param when using FileStream::from_path
`FileStream::from_path()` already knows its return type but because it was in the generic impl for
`S: TryStream + Send + 'static`, it had to be called with an exact type, basically requiring to add
the `tokio_util` crate to direct dependencies to use it, losing helf the convenience of the method.

I added a comment to explain why this a separate impl to future readers and modified the doc example
but everything else is just moving code around.