Add customizable WebSocket subprotocol selection.
This adds a getter for the Sec-WebSocket-Protocol requested by the
client and a setter for the chosen subprotocol, which enables more
flexible subprotocol selection logic, such as token-encoded
subprotocols or runtime-info-based selection.

Token-encoded subprotocols are a common practice for authentication in
WebSocket. Although this is not a standard usage, it is used in
practice, for example by Kubernetes:
https://github.com/kubernetes/kubernetes/commit/714f97d7baf4975ad3aa47735a868a81a984d1f0